WIP: vl53lxx cleanup
